refactor: modularize interpreter
