+++ interpreter/interpreter.go
@@ -8,8 +8,7 @@ import (
 	"github.com/alecthomas/participle/v2/lexer"
 	"github.com/fatih/color"
 	"reflect"
-	"sort"
-	"strconv"
+	"strings"
 	"sync"
 
 	"mochi/parser"
@@ -20,206 +19,8 @@ import (
 	"mochi/runtime/stream"
 	"mochi/types"
 	"os"
-	"strings"
 	"time"
 )
-
-// ValueTag represents the type of a runtime value.
-type ValueTag uint8
-
-const (
-	TagNull ValueTag = iota
-	TagInt
-	TagFloat
-	TagStr
-	TagBool
-	TagList
-	TagMap
-	TagFunc
-)
-
-func (t ValueTag) String() string {
-	switch t {
-	case TagNull:
-		return "null"
-	case TagInt:
-		return "int"
-	case TagFloat:
-		return "float"
-	case TagStr:
-		return "string"
-	case TagBool:
-		return "bool"
-	case TagList:
-		return "list"
-	case TagMap:
-		return "map"
-	case TagFunc:
-		return "func"
-	default:
-		return "unknown"
-	}
-}
-
-// Value is a tagged union used at runtime to avoid reflection.
-type Value struct {
-	Tag   ValueTag
-	Int   int
-	Float float64
-	Str   string
-	Bool  bool
-	List  []Value
-	Map   map[string]Value
-	Func  *closure
-}
-
-// Truthy returns the boolean interpretation of v.
-func (v Value) Truthy() bool {
-	switch v.Tag {
-	case TagBool:
-		return v.Bool
-	case TagInt:
-		return v.Int != 0
-	case TagFloat:
-		return v.Float != 0
-	case TagStr:
-		return v.Str != ""
-	case TagList:
-		return len(v.List) > 0
-	case TagMap:
-		return len(v.Map) > 0
-	default:
-		return false
-	}
-}
-
-func anyToValue(v any) Value {
-	switch val := v.(type) {
-	case nil:
-		return Value{Tag: TagNull}
-	case int:
-		return Value{Tag: TagInt, Int: val}
-	case int64:
-		return Value{Tag: TagInt, Int: int(val)}
-	case float64:
-		return Value{Tag: TagFloat, Float: val}
-	case string:
-		return Value{Tag: TagStr, Str: val}
-	case bool:
-		return Value{Tag: TagBool, Bool: val}
-	case []Value:
-		return Value{Tag: TagList, List: val}
-	case []any:
-		list := make([]Value, len(val))
-		for i, x := range val {
-			list[i] = anyToValue(x)
-		}
-		return Value{Tag: TagList, List: list}
-	case map[string]Value:
-		return Value{Tag: TagMap, Map: val}
-	case map[string]any:
-		m := make(map[string]Value, len(val))
-		for k, x := range val {
-			m[k] = anyToValue(x)
-		}
-		return Value{Tag: TagMap, Map: m}
-	case closure:
-		return Value{Tag: TagFunc, Func: &val}
-	case *closure:
-		return Value{Tag: TagFunc, Func: val}
-	default:
-		return Value{Tag: TagNull}
-	}
-}
-
-func valueToAny(v Value) any {
-	switch v.Tag {
-	case TagInt:
-		return v.Int
-	case TagFloat:
-		return v.Float
-	case TagStr:
-		return v.Str
-	case TagBool:
-		return v.Bool
-	case TagList:
-		list := make([]any, len(v.List))
-		for i, x := range v.List {
-			list[i] = valueToAny(x)
-		}
-		return list
-	case TagMap:
-		m := make(map[string]any, len(v.Map))
-		for k, x := range v.Map {
-			m[k] = valueToAny(x)
-		}
-		return m
-	case TagFunc:
-		return v.Func
-	default:
-		return nil
-	}
-}
-
-func argsKey(args []any) string {
-	var sb strings.Builder
-	for i, a := range args {
-		if i > 0 {
-			sb.WriteByte('|')
-		}
-		writeArgKey(&sb, a)
-	}
-	return sb.String()
-}
-
-func writeArgKey(sb *strings.Builder, v any) {
-	switch x := v.(type) {
-	case int:
-		sb.WriteString("i:")
-		sb.WriteString(strconv.Itoa(x))
-	case float64:
-		sb.WriteString("f:")
-		sb.WriteString(strconv.FormatFloat(x, 'g', -1, 64))
-	case string:
-		sb.WriteString("s:")
-		sb.WriteString(strconv.Quote(x))
-	case bool:
-		sb.WriteString("b:")
-		if x {
-			sb.WriteByte('1')
-		} else {
-			sb.WriteByte('0')
-		}
-	case []any:
-		sb.WriteByte('[')
-		for i, e := range x {
-			if i > 0 {
-				sb.WriteByte(',')
-			}
-			writeArgKey(sb, e)
-		}
-		sb.WriteByte(']')
-	case map[string]any:
-		sb.WriteByte('{')
-		keys := make([]string, 0, len(x))
-		for k := range x {
-			keys = append(keys, k)
-		}
-		sort.Strings(keys)
-		for i, k := range keys {
-			if i > 0 {
-				sb.WriteByte(',')
-			}
-			sb.WriteString(k)
-			sb.WriteByte(':')
-			writeArgKey(sb, x[k])
-		}
-		sb.WriteByte('}')
-	default:
-		sb.WriteString("p:")
-		sb.WriteString(fmt.Sprintf("%p", v))
-	}
-}
 
 // Interpreter executes Mochi programs using a shared runtime and type environment.
 type Interpreter struct {
@@ -1795,322 +1596,6 @@ func (i *Interpreter) evalLiteral(l *parser.Literal) (any, error) {
 		return bool(*l.Bool), nil
 	default:
 		return nil, errInvalidLiteral(l.Pos)
-	}
-}
-
-func builtinPrint(i *Interpreter, c *parser.CallExpr) (any, error) {
-	var sb strings.Builder
-	for _, arg := range c.Args {
-		val, err := i.evalExpr(arg)
-		if err != nil {
-			return nil, err
-		}
-		fmt.Fprintf(&sb, "%v ", val)
-	}
-	_, err := fmt.Fprintln(i.env.Writer(), strings.TrimSpace(sb.String()))
-	return nil, err
-}
-
-func builtinLen(i *Interpreter, c *parser.CallExpr) (any, error) {
-	if len(c.Args) != 1 {
-		return nil, errTooManyFunctionArgs(c.Pos, "len", 1, len(c.Args))
-	}
-	val, err := i.evalExpr(c.Args[0])
-	if err != nil {
-		return nil, err
-	}
-
-	switch v := val.(type) {
-	case []any:
-		return len(v), nil
-	case string:
-		return len([]rune(v)), nil
-	case map[string]any:
-		return len(v), nil
-	default:
-		return nil, errInvalidLenOperand(c.Pos, fmt.Sprintf("%T", val))
-	}
-}
-
-func builtinNow(i *Interpreter, c *parser.CallExpr) (any, error) {
-	if len(c.Args) != 0 {
-		return nil, fmt.Errorf("now() takes no arguments")
-	}
-	return time.Now().UnixNano(), nil
-}
-
-func builtinJSON(i *Interpreter, c *parser.CallExpr) (any, error) {
-	if len(c.Args) != 1 {
-		return nil, fmt.Errorf("json(x) takes exactly one argument")
-	}
-	val, err := i.evalExpr(c.Args[0])
-	if err != nil {
-		return nil, err
-	}
-	data, err := json.MarshalIndent(val, "", "  ")
-	if err != nil {
-		return nil, err
-	}
-	_, err = fmt.Fprintln(i.env.Writer(), string(data))
-	return nil, err
-}
-
-func builtinStr(i *Interpreter, c *parser.CallExpr) (any, error) {
-	if len(c.Args) != 1 {
-		return nil, fmt.Errorf("str(x) takes exactly one argument")
-	}
-	val, err := i.evalExpr(c.Args[0])
-	if err != nil {
-		return nil, err
-	}
-	return fmt.Sprint(val), nil
-}
-
-func builtinSleep(i *Interpreter, c *parser.CallExpr) (any, error) {
-	if len(c.Args) != 1 {
-		return nil, fmt.Errorf("sleep(x) takes exactly one argument")
-	}
-	val, err := i.evalExpr(c.Args[0])
-	if err != nil {
-		return nil, err
-	}
-	switch v := val.(type) {
-	case int:
-		time.Sleep(time.Duration(v) * time.Millisecond)
-	case int64:
-		time.Sleep(time.Duration(v) * time.Millisecond)
-	case float64:
-		time.Sleep(time.Duration(v * float64(time.Millisecond)))
-	}
-	return nil, nil
-}
-
-func builtinCount(i *Interpreter, c *parser.CallExpr) (any, error) {
-	if len(c.Args) != 1 {
-		return nil, fmt.Errorf("count(x) takes exactly one argument")
-	}
-	val, err := i.evalExpr(c.Args[0])
-	if err != nil {
-		return nil, err
-	}
-	switch v := val.(type) {
-	case []any:
-		return len(v), nil
-	case *data.Group:
-		return len(v.Items), nil
-	default:
-		return nil, fmt.Errorf("count() expects list or group, got %T", val)
-	}
-}
-
-func builtinAvg(i *Interpreter, c *parser.CallExpr) (any, error) {
-	if len(c.Args) != 1 {
-		return nil, fmt.Errorf("avg(x) takes exactly one argument")
-	}
-	val, err := i.evalExpr(c.Args[0])
-	if err != nil {
-		return nil, err
-	}
-	var list []any
-	switch v := val.(type) {
-	case []any:
-		list = v
-	case *data.Group:
-		list = v.Items
-	default:
-		return nil, fmt.Errorf("avg() expects list or group, got %T", val)
-	}
-	if len(list) == 0 {
-		return 0, nil
-	}
-	var sum float64
-	for _, it := range list {
-		switch n := it.(type) {
-		case int:
-			sum += float64(n)
-		case int64:
-			sum += float64(n)
-		case float64:
-			sum += n
-		default:
-			return nil, fmt.Errorf("avg() expects numbers, got %T", it)
-		}
-	}
-	return sum / float64(len(list)), nil
-}
-
-func (i *Interpreter) invokeTool(cl closure, args map[string]any) (any, error) {
-	child := types.NewEnv(cl.Env)
-	for _, param := range cl.FullParams {
-		child.SetValue(param.Name, args[param.Name], true)
-	}
-	old := i.env
-	i.env = child
-	defer func() { i.env = old }()
-	if cl.Fn.ExprBody != nil {
-		return i.evalExpr(cl.Fn.ExprBody)
-	}
-	for _, stmt := range cl.Fn.BlockBody {
-		if err := i.evalStmt(stmt); err != nil {
-			if r, ok := err.(returnSignal); ok {
-				return r.value, nil
-			}
-			return nil, err
-		}
-	}
-	return nil, nil
-}
-
-func isUnderscoreExpr(e *parser.Expr) bool {
-	if e == nil {
-		return false
-	}
-	if len(e.Binary.Right) != 0 {
-		return false
-	}
-	u := e.Binary.Left
-	if len(u.Ops) != 0 {
-		return false
-	}
-	p := u.Value
-	if len(p.Ops) != 0 {
-		return false
-	}
-	if p.Target.Selector != nil && p.Target.Selector.Root == "_" && len(p.Target.Selector.Tail) == 0 {
-		return true
-	}
-	return false
-}
-
-func identName(e *parser.Expr) (string, bool) {
-	if e == nil {
-		return "", false
-	}
-	if len(e.Binary.Right) != 0 {
-		return "", false
-	}
-	u := e.Binary.Left
-	if len(u.Ops) != 0 {
-		return "", false
-	}
-	p := u.Value
-	if len(p.Ops) != 0 {
-		return "", false
-	}
-	if p.Target.Selector != nil && len(p.Target.Selector.Tail) == 0 {
-		return p.Target.Selector.Root, true
-	}
-	return "", false
-}
-
-func callPattern(e *parser.Expr) (*parser.CallExpr, bool) {
-	if e == nil {
-		return nil, false
-	}
-	if len(e.Binary.Right) != 0 {
-		return nil, false
-	}
-	u := e.Binary.Left
-	if len(u.Ops) != 0 {
-		return nil, false
-	}
-	p := u.Value
-	if len(p.Ops) != 0 || p.Target.Call == nil {
-		return nil, false
-	}
-	return p.Target.Call, true
-}
-
-func (i *Interpreter) evalQuery(q *parser.QueryExpr) (any, error) {
-	src, err := i.evalExpr(q.Source)
-	if err != nil {
-		return nil, err
-	}
-
-	child := types.NewEnv(i.env)
-	old := i.env
-	i.env = child
-	defer func() { i.env = old }()
-
-	return data.EvalQuery(src, q, child, func(e *parser.Expr) (any, error) {
-		return i.evalExpr(e)
-	})
-}
-
-func (i *Interpreter) evalMatch(m *parser.MatchExpr) (any, error) {
-	val, err := i.evalExpr(m.Target)
-	if err != nil {
-		return nil, err
-	}
-	for _, c := range m.Cases {
-		if isUnderscoreExpr(c.Pattern) {
-			return i.evalExpr(c.Result)
-		}
-		// Variant pattern like Node(left, value, right)
-		if call, ok := callPattern(c.Pattern); ok {
-			obj, ok := val.(map[string]any)
-			if !ok {
-				continue
-			}
-			name, _ := obj["__name"].(string)
-			if name != call.Func {
-				continue
-			}
-			st, ok := i.types.GetStruct(call.Func)
-			if !ok {
-				continue
-			}
-			if len(call.Args) != len(st.Order) {
-				continue
-			}
-			child := types.NewEnv(i.env)
-			for idx, arg := range call.Args {
-				if n, ok := identName(arg); ok {
-					child.SetValue(n, obj[st.Order[idx]], true)
-				}
-			}
-			old := i.env
-			i.env = child
-			res, err := i.evalExpr(c.Result)
-			i.env = old
-			return res, err
-		}
-
-		if ident, ok := identName(c.Pattern); ok {
-			obj, ok := val.(map[string]any)
-			if ok {
-				name, _ := obj["__name"].(string)
-				if name == ident {
-					return i.evalExpr(c.Result)
-				}
-			}
-		} else {
-			pv, err := i.evalExpr(c.Pattern)
-			if err != nil {
-				return nil, err
-			}
-			eq, err := applyBinary(c.Pos, val, "==", pv)
-			if err != nil {
-				return nil, err
-			}
-			if b, ok := eq.(bool); ok && b {
-				return i.evalExpr(c.Result)
-			}
-		}
-	}
-	return nil, nil
-}
-
-func (i *Interpreter) builtinFuncs() map[string]func(*Interpreter, *parser.CallExpr) (any, error) {
-	return map[string]func(*Interpreter, *parser.CallExpr) (any, error){
-		"print": builtinPrint,
-		"len":   builtinLen,
-		"now":   builtinNow,
-		"json":  builtinJSON,
-		"str":   builtinStr,
-		"sleep": builtinSleep,
-		"count": builtinCount,
-		"avg":   builtinAvg,
 	}
 }
